Timing improved with a price
Switching from sleeps with every MIDI event to batching them up
definitely helps, but short sleeps don't work so well in Python. Using
busy-waits for now, but that's not a good long-term solution. Also,
there is a weird bug that I've traced into the play() call for the first
few notes. I'm thinking that the first few requests get bogged down,
somehow. Not sure if the problem is on the host or the client.